+++ tests/test_pm.py
@@ -4,10 +4,10 @@ import os
 import sys
 import time
 import unittest
-from flask import Flask, g, request
-from flask_pluginkit import Flask as ExFlask, PluginManager, LocalStorage
-from flask_pluginkit.exceptions import PluginError
-from flask_pluginkit._compat import PY2
+from flask import Flask, request, Markup
+from flask_pluginkit import Flask as ExFlask, PluginManager, LocalStorage, push_dcp
+from flask_pluginkit.exceptions import PluginError, NotCallableError
+from flask_pluginkit._compat import PY2, iteritems
 from jinja2 import ChoiceLoader
 EXAMPLE_DIR = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), 'examples/fulldemo')
 sys.path.append(EXAMPLE_DIR)
@@ -44,7 +44,6 @@ class PMTest(unittest.TestCase):
         self.app1_pm = app1.extensions.get("pluginkit")
         self.app3_pm = app3.extensions.get("pluginkit")
         self.app4_pm = app4.extensions.get("pluginkit")
-        self.app4_client = app4.test_client()
 
     def test_extself(self):
         self.assertIsInstance(self.app1_pm, PluginManager)
@@ -100,7 +99,7 @@ class PMTest(unittest.TestCase):
                 self.assertTrue("js/hello.js" in data)
                 self.assertTrue(self.app4_pm.pluginkit_config['whoami'] in data)
         self.assertEqual(len(app4.blueprints), 2)
-        self.assertEqual(len(self.app4_pm.get_all_plugins), 2)
+        self.assertEqual(len(self.app4_pm.get_all_plugins), 3)
         self.assertEqual(len(self.app4_pm.get_enabled_beps), 2)
         self.assertIn(self.app4_pm.static_endpoint, app4.view_functions)
         with app4.test_request_context():
@@ -114,6 +113,51 @@ class PMTest(unittest.TestCase):
         self.assertEqual(4, len(view_limit_data))
         with app4.test_request_context('/limit/test'):
             self.assertEqual("view_limit", request.endpoint)
+
+    def test_filter(self):
+        self.assertIn("repeat_filter", app4.jinja_env.filters)
+        self.assertIn("demo_filter2", app4.jinja_env.filters)
+        self.assertEqual('test-filter-repeat', app4.jinja_env.filters['repeat_filter']())
+        self.assertEqual('test-filter', app4.jinja_env.filters['demo_filter2']())
+
+    def test_errhandler(self):
+        ehs = app4.error_handler_spec[None]
+        self.assertIsInstance(ehs, dict)
+        self.assertIn(403, ehs)
+        self.assertIn(404, ehs)
+        with app4.test_client() as c:
+            resp404 = c.get('/404')
+            self.assertEqual(404, resp404.status_code)
+            self.assertIn(b"Not Found Page", resp404.data)
+            resp403 = c.get('/403')
+            self.assertEqual(403, resp403.status_code)
+            self.assertIn(b"permission deny", resp403.data)
+
+    def test_tcp(self):
+        context = {
+            k: v
+            for i in app4.template_context_processors[None]
+            for k, v in iteritems(i())
+        }
+        self.assertIsInstance(context, dict)
+        self.assertIn("timestamp", context)
+        self.assertIn("change_to_str", context)
+        self.assertTrue(context['change_to_str'], str)
+
+    def test_dcp(self):
+        def callback():
+            return "test"
+        self.assertRaises(AttributeError, push_dcp, 'raise', callback)
+        with app4.app_context():
+            push_dcp("test", callback)
+            result = [f() for f in self.app4_pm._dcp_manager.list["test"]]
+            ft = self.app4_pm._dcp_manager.emit("test")
+            self.assertRaises(PluginError, push_dcp, ['raise'], callback)
+            self.assertRaises(NotCallableError, push_dcp, 'raise', "abc")
+            self.assertIsInstance(ft, Markup)
+            self.assertEqual(ft, Markup("test"))
+            self.assertEqual(result, ["test"])
+
 
 if __name__ == '__main__':
     unittest.main()