+++ CHANGELOG.rst
@@ -1,5 +1,21 @@
 .. currentmodule:: flask-pluginkit
 
+v3.2.0
+------
+
+Released in 2019-07-24
+
+- feat: dcp, be similar to flask-plugins event system,
+        add :class:`~flask_pluginkit.utils.DcpManager`
+        and :func:`~flask_pluginkit.push_dcp`
+- feat: add error handler
+- feat: add template filter
+- feat: add template context processor
+- test: add test case
+- docs: update docs
+- chore: update example
+- chore: make some of the code more elegant
+
 v3.1.0
 ------
 
@@ -17,7 +33,6 @@ v3.0.1
 Released in 2019-07-21
 
 - The register functin can return empty dict, not necessarily meaningless.
-
 - Update docs and translate and add test.
 
 v3.0.0